chore: move getNetworkNonce and handleNetworkSwitch
