+++ app/core/DeeplinkManager/Handlers/switchNetwork.test.ts
@@ -2,7 +2,7 @@ import switchNetwork from './switchNetwork';
 import DeeplinkManager from '../DeeplinkManager';
 import { showAlert } from '../../../actions/alert';
 import { strings } from '../../../../locales/i18n';
-import { handleNetworkSwitch } from '../../../util/networks';
+import handleNetworkSwitch from '../../../util/networks/handleNetworkSwitch';
 
 jest.mock('../../../util/networks', () => ({
   handleNetworkSwitch: jest.fn(),
