+++ app/util/networks/index.test.ts
@@ -8,7 +8,6 @@ import {
   compareRpcUrls,
   getBlockExplorerAddressUrl,
   getBlockExplorerTxUrl,
-  getNetworkNonce,
 } from '.';
 import {
   MAINNET,
@@ -19,7 +18,6 @@ import {
   LINEA_MAINNET,
 } from '../../../app/constants/network';
 import { NetworkSwitchErrorType } from '../../../app/constants/error';
-import Engine from '../../core/Engine';
 
 jest.mock('./../../core/Engine', () => ({
   context: {
@@ -278,36 +276,6 @@ describe('network-utils', () => {
 
       expect(url).toBe(`https://lineascan.build/tx/${mockTransactionHash}`);
       expect(title).toBe(`lineascan.build`);
-    });
-  });
-
-  describe('getNetworkNonce', () => {
-    const nonceMock = 123;
-    const fromMock = '0x123';
-
-    it('returns value from TransactionController', async () => {
-      Engine.context.TransactionController.getNonceLock.mockReturnValueOnce({
-        nextNonce: nonceMock,
-        releaseLock: jest.fn(),
-      });
-
-      expect(await getNetworkNonce({ from: fromMock })).toBe(nonceMock);
-
-      expect(
-        Engine.context.TransactionController.getNonceLock,
-      ).toHaveBeenCalledWith(fromMock);
-    });
-
-    it('releases nonce lock', async () => {
-      const releaseLockMock = jest.fn();
-
-      Engine.context.TransactionController.getNonceLock.mockReturnValueOnce({
-        releaseLock: releaseLockMock,
-      });
-
-      await getNetworkNonce({ from: fromMock });
-
-      expect(releaseLockMock).toHaveBeenCalledTimes(1);
     });
   });
 });
